fix: 유효성 검사가 나오지 않아 해당 부분 수정
- useRefs 제거
- useValidation에 합침 의존성으로 인해
